apps/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlmodel==0.0.14
asyncpg==0.29.0
python-multipart==0.0.6
pydantic-settings==2.1.0
httpx==0.26.0
boto3==1.34.0
slowapi==0.1.9
